refactor: all files and folder restructuring for the app
