+++ src/index.ts
@@ -1,33 +1,24 @@
-import { LogsUtil } from './utils/logs-util';
-import { AppProcess } from './app/app-process';
-import { apiConfig } from './configs/api-config';
-import { AppUtil } from './utils/app-util';
+import { LogsUtil, AppUtil } from './utils';
 
 starApp();
 
 async function starApp() {
+  const logsUtil = new LogsUtil();
+  await logsUtil.clearLogs();
+  await logsUtil.addLogs('info', 'start an app', 'app');
   try {
     const { fromDueDateLimit, toDueDateLimit, workingDays } =
       AppUtil.getStartEndDateLimit();
-    const appProcess = new AppProcess(apiConfig, workingDays);
-    const logsUtil = new LogsUtil();
-    await logsUtil.clearLogs();
-    await logsUtil.addLogs('info', 'start an app', 'app');
-    await appProcess.setWorkspaceFolders();
-    await appProcess.setWorkSpaceUsers();
-    await appProcess.generateWorkSpaceFolderReport();
-    await appProcess.generateWorkSpaceUserReport();
-    await appProcess.setAllTask(fromDueDateLimit, toDueDateLimit);
-    await appProcess.generateSourceReportFile();
-    await appProcess.generateTaskSummary(fromDueDateLimit, toDueDateLimit);
-    await appProcess.generatePayrollForStaff();
-    await appProcess.generateTimeSheetForIndividual(
+    console.log({
       fromDueDateLimit,
-      toDueDateLimit
-    );
-    await logsUtil.addLogs('info', 'End of script', 'app');
+      toDueDateLimit,
+      workingDays,
+      fromDate: AppUtil.getFormattedDate(fromDueDateLimit),
+      toDate: AppUtil.getFormattedDate(toDueDateLimit)
+    });
   } catch (error: any) {
     error = error.message || error;
     console.log({ error });
   }
+  await logsUtil.addLogs('info', 'End of script', 'app');
 }
